Move logging to separate file
