+++ src/inventa/utils/utils.go
@@ -4,8 +4,7 @@ import (
 	"os"
 	"regexp"
 
-	"github.com/osrg/gobgp/v3/pkg/log"
-	"github.com/sirupsen/logrus"
+	"github.com/shelson/inventa/src/inventa/logging"
 	"gopkg.in/yaml.v2"
 )
 
@@ -22,9 +21,6 @@ type Conf struct {
 	GroupSplitChar        string   `yaml:"group_split_char"`
 	GroupSplitIndex       int      `yaml:"group_split_index"`
 }
-
-// Log is the logging object
-var Log *logrus.Logger
 
 // Configs is our shared config object
 var Configs Conf
@@ -43,19 +39,12 @@ func FindInArray(what string, where []string) (idx int, found bool) {
 func InitConfig() {
 	yamlFile, err := os.ReadFile("config.yaml")
 	if err != nil {
-		Log.Fatal(err)
+		logging.Log.Fatal(err)
 	}
 	err = yaml.Unmarshal(yamlFile, &Configs)
 	if err != nil {
-		Log.Fatal(err)
+		logging.Log.Fatal(err)
 	}
-}
-
-// SetUpLogger sets up the logger variable
-func SetUpLogger() {
-	Log = logrus.New()
-	Log.SetLevel(logrus.InfoLevel)
-	Log.Info("Set Logger Up")
 }
 
 // StripUnwanted removes any substrings from our name string
@@ -63,54 +52,9 @@ func StripUnwanted(name string) string {
 	for _, pattern := range Configs.NodeNameStripPatterns {
 		re, err := regexp.Compile(pattern)
 		if err != nil {
-			Log.Fatal(err)
+			logging.Log.Fatal(err)
 		}
 		name = re.ReplaceAllString(name, "")
 	}
 	return name
-}
-
-// MyLogger implements github.com/osrg/gobgp/v3/pkg/log/Logger interface
-type MyLogger struct {
-	Logger *logrus.Logger
-}
-
-// Panic level
-func (l *MyLogger) Panic(msg string, fields log.Fields) {
-	l.Logger.WithFields(logrus.Fields(fields)).Panic(msg)
-}
-
-// Fatal Level
-func (l *MyLogger) Fatal(msg string, fields log.Fields) {
-	l.Logger.WithFields(logrus.Fields(fields)).Fatal(msg)
-}
-
-// Error level
-func (l *MyLogger) Error(msg string, fields log.Fields) {
-	l.Logger.WithFields(logrus.Fields(fields)).Error(msg)
-}
-
-// Warn level
-func (l *MyLogger) Warn(msg string, fields log.Fields) {
-	l.Logger.WithFields(logrus.Fields(fields)).Warn(msg)
-}
-
-// Info level
-func (l *MyLogger) Info(msg string, fields log.Fields) {
-	l.Logger.WithFields(logrus.Fields(fields)).Info(msg)
-}
-
-// Debug level
-func (l *MyLogger) Debug(msg string, fields log.Fields) {
-	l.Logger.WithFields(logrus.Fields(fields)).Debug(msg)
-}
-
-// SetLevel sets the level
-func (l *MyLogger) SetLevel(level log.LogLevel) {
-	l.Logger.SetLevel(logrus.Level(level))
-}
-
-// GetLevel gets the level
-func (l *MyLogger) GetLevel() log.LogLevel {
-	return log.LogLevel(l.Logger.GetLevel())
 }
